recruitment-helper/requirements.txt
gradio==4.19.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dateutil==2.8.2
python-dotenv==1.0.0 